Solve circular import differently
